Split out converters in separate file.
